apps/ai-service/requirements.txt
fastapi
uvicorn
pydantic
pytest
pytest-html
httpx
python-dotenv
python-jose[cryptography]
authlib
# optional: transformers

